[vpr] move files to corresponding folder in librrgraph/src
