+++ libs/librrgraph/src/utils/alloc_and_load_rr_indexed_data.h
