+++ libs/librrgraph/src/utils/describe_rr_node.h
