+++ vpr/src/route/rr_graph.h
@@ -8,7 +8,7 @@
 
 #include "device_grid.h"
 #include "vpr_types.h"
-#include "graph_type.h"
+#include "rr_graph_type.h"
 #include "describe_rr_node.h"
 
 /* Warnings about the routing graph that can be returned.
